requirements.txt
discord.py
colorama
pillow
python-dotenv
requests